requirements.txt
asgiref~=3.5.0
certifi~=2021.10.8
charset-normalizer~=2.0.11
Django~=4.0.1
django-modeladmin-reorder~=0.3.1
django-preferences~=1.0.0
idna~=3.3
pyTelegramBotAPI~=4.4.0
pytz~=2021.3
requests~=2.27.1
sqlparse~=0.4.1
urllib3~=1.26.8
